multi-modal-retrieval-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic-core~=2.27.2
pydantic-settings~=2.7.1
python-dotenv>=1.0.0
feast==0.45.0
utils==1.0.2
transformers~=4.48.3
sentence-transformers~=3.4.1
faiss-cpu~=1.9.0
fastparquet~=2024.5.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.26.0
numpy~=1.26.4
pydantic~=2.10.6
pillow~=11.1.0
